I did a bit of work off camera...
I put the css and js files in folders for organization. The style work for the left colum is in "podcast_shows.css", the middle is in "shows.css", the left is in "draggableStyle.css", and the general stuff is in "style.css". I also added the functionality for the middle section to display all the epidodes instead of just one.